+++ packages/jest-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @batoanng/jest-config
 
+## 1.4.1
+
+### Patch Changes
+
+- Minimize package size
+
 ## 1.4.0
 
 ### Minor Changes