+++ packages/prettier-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # prettier-config
 
+## 1.6.1
+
+### Patch Changes
+
+- Minimize package size
+
 ## 1.6.0
 
 ### Minor Changes