+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @batoanng/utils
 
+## 2.4.1
+
+### Patch Changes
+
+- Minimize package size
+
 ## 2.4.0
 
 ### Minor Changes